Little adjustments in the folders set
