+++ DataStructures/Tree/BinaryTree.java
@@ -1,4 +1,4 @@
-package Tree;
+package DataStructures.Tree;
 
 public class BinaryTree {
